+++ packages/template-retail-react-app/config/mocks/default.js
@@ -35,11 +35,15 @@ module.exports = {
         login: {
             passwordless: {
                 enabled: false,
-                callbackURI: 'https://webhook.site/27761b71-50c1-4097-a600-21a3b89a546c'
+                callbackURI: 'https://webhook.site/27761b71-50c1-4097-a600-21a3b89a546c',
+                landingPath: '/passwordless-login-landing'
             },
             social: {
                 enabled: false,
                 idps: ['google', 'apple']
+            },
+            resetPassword: {
+                landingPath: '/reset-password-landing'
             }
         },
         siteAliases: {